+++ CHANGELOG.md
@@ -542,3 +542,4 @@
 * Implemented account recovery in case of deleted accounts, both generated and imported
 * Added test cases for account recovery
 * Added logs for export mnemonic
+* Added logs for export private key(for generated and imported accounts)