+++ CHANGELOG.md
@@ -543,3 +543,4 @@
 * Added test cases for account recovery
 * Added logs for export mnemonic
 * Added logs for export private key(for generated and imported accounts)
+* Re-encrypted the private key of imported wallets and updated vault state after change pin